+++ sdks/typescript/react/src/components/full/index.ts
@@ -1,5 +1,7 @@
 export * from "./EndScreen";
 export * from "./GuidedTour";
 export * from "./LoadingScreen";
+export * from "./PreJoinLobby";
+export * from "./SessionMeetingRoom";
 export * from "./SplitStage";
 export * from "./WhiteboardPanel";